# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GoAI Multi-Agent Runtime API
  version: 1.0.0
  description: >-
    GoAI is a multi-Agent protocol runtime. AG-UI connects clients to a runtime,
    A2A connects Agents to Agents, and Workflow/Eino Graph describes an Agent's
    execution capability. Kafka is an internal scheduling mechanism and is not
    an Agent communication protocol.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          $ref: '#/components/responses/Envelope'
        '400':
          $ref: '#/components/responses/Error'
  /auth/login:
    post:
      tags: [Auth]
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          $ref: '#/components/responses/Envelope'
        '401':
          $ref: '#/components/responses/Error'
...
  /api/users:
    post:
      tags: [Users]
      security: [{bearerAuth: []}]
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          $ref: '#/components/responses/Envelope'
    get:
      tags: [Users]
      security: [{bearerAuth: []}]
      operationId: listUsers
      responses:
        '200':
          $ref: '#/components/responses/Envelope'
  /api/users/{id}:
    get:
      tags: [Users]
      security: [{bearerAuth: []}]
      operationId: getUser
      parameters: [{ $ref: '#/components/parameters/UserID' }]
      responses:
        '200':
          $ref: '#/components/responses/Envelope'
    put:
      tags: [Users]
      security: [{bearerAuth: []}]
      operationId: updateUser
      parameters: [{ $ref: '#/components/parameters/UserID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                email: {type: string, format: email}
                password: {type: string, format: password}
      responses:
        '200':
          $ref: '#/components/responses/Envelope'
    delete:
      tags: [Users]
      security: [{bearerAuth: []}]
      operationId: deleteUser
      parameters: [{ $ref: '#/components/parameters/UserID' }]
      responses:
        '200':
          $ref: '#/components/responses/Envelope'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    a2aHMAC:
      type: apiKey
      in: header
      name: Authorization
      description: 'Use the GoAI-HMAC-SHA256 scheme. Companion headers are defined below.'
  parameters:
...
    UserID:
      name: id
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
  responses:
    Envelope:
      description: Unified JSON response envelope.
      headers:
        X-Trace-ID:
          schema: {type: string}
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
    Error:
      description: Unified error envelope.
      headers:
        X-Trace-ID:
          schema: {type: string}
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
  schemas:
    Envelope:
      type: object
      required: [code, message, data, trace_id]
      properties:
        code: {type: string, example: OK}
        message: {type: string, example: success}
        data: {}
        trace_id: {type: string, example: trace_xxx}
    RegisterRequest:
      type: object
      required: [username, email, password]
      additionalProperties: false
      properties:
        username: {type: string, minLength: 3, maxLength: 64}
        email: {type: string, format: email}
        password: {type: string, format: password, minLength: 8}
    LoginRequest:
      type: object
      required: [username, password]
      additionalProperties: false
      properties:
        username: {type: string}
        password: {type: string, format: password}
...